# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Daycore v2 API
  version: "2.4.0"
  description: >
    Daycore v2 backend (Go). Autonomous AI day-planning + mood companion.
    Anonymous-first: POST /api/session/init issues an httpOnly, signed session
    cookie (dc_sid) that authorizes all data endpoints. Optional email/OAuth
    sign-in sets an httpOnly JWT cookie (dc_auth); logout revokes all issued JWTs
    (server-side token version). Same-origin web frontends MUST send cookies
    (fetch: credentials:"include"). Native apps and separately-deployed frontends
    use header credentials instead (dual-rail, cookies stay supported): the
    signed session token in X-Session-Token (returned by /api/session/init with
    {"tokenInBody":true}) and/or the JWT in Authorization: Bearer (returned by
    login/register with {"tokenInBody":true}). Check GET /api/version (apiVersion
    contract major) before use — matching rules in FRONTEND_HANDOFF.md.
    The browser extension authenticates to /api/import/* with X-Import-Token;
    admin endpoints use X-Admin-Token.
    Errors are {error, message}. The single SSE endpoint is /api/ai/companion —
    see FRONTEND_HANDOFF.md §B for the v2 frame protocol (NOT the legacy [DONE]).
servers:
  - url: http://localhost:8080
    description: Local (Go serves the API; nginx proxies /api in production)
...
security:
  - sessionCookie: []
  - sessionToken: []
  - bearerAuth: []
paths:
...
  /api/v2/ai/companion:
    post:
      operationId: companion
      tags: [ai]
      summary: Companion agent over SSE v2 (tool-calls frame protocol).
      description: >
        Returns text/event-stream. Each frame is `data: {json}\n\n` discriminated
        by `type`: delta | reasoning | tool_start | tool_result | decision_card |
        error | done. Comment heartbeats `: ping`. NO `[DONE]` sentinel. See
        FRONTEND_HANDOFF.md §B. With threadId the server loads AND persists thread
        history; client roles are whitelisted to user/assistant.
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { message: { type: string, description: "May be empty when attachmentIds is not — dropping in a photo with no words is a complete thing to say." }, timezone: { type: string }, location: { type: string, maxLength: 128, description: "Where the device is, free text. Fills in or updates a device-detected location; never overwrites one the user set on the settings page. Used only by the briefs — the chat tool takes a location as an argument." }, assistantName: { type: string }, threadId: { type: string }, attachmentIds: { type: array, maxItems: 20, items: { type: string }, description: "Ids from POST /api/files. Must be this session's unsent uploads; they become part of the persisted user message. Files the chosen model cannot read are described to it by name rather than dropped." }, conversationHistory: { type: array, items: { type: object, properties: { role: { type: string, enum: [user, assistant] }, content: { type: string } } } } } } } } }
      responses: { '200': { description: SSE stream (text/event-stream), content: { text/event-stream: { schema: { type: string }, example: 'data: {"type":"delta","text":"好的"}\n\ndata: {"type":"done"}\n\n' } } } }
...
  /api/v2/admin/stats:
    get: { operationId: adminStats, tags: [admin], security: [{ adminToken: [] }], summary: Dashboard aggregation (incl. feedbackUseful/feedbackTotal)., responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AdminStats' } } } } } }
  /api/v2/admin/ailogs:
    get:
      operationId: adminAILogs
      tags: [admin]
      security: [{ adminToken: [] }]
      summary: The AI call ledger, newest first.
      description: >-
        One row per model call: endpoint, model, token counts, duration, status.

        ⚠️ NO REQUEST OR RESPONSE BODIES, and that is a boundary rather than a gap.
        The request body of a companion call IS the user's conversation, and
        ailogs.read promises the opposite ("今天不含对话正文"). If a capture is ever
        added it belongs behind db.user_content or a permission of its own, never
        behind this one.

        Paging is KEYSET, not offset: send `beforeAt` and `beforeId` from the
        previous page's last row (the response hands them back as `nextBeforeAt` /
        `nextBeforeId`). Both halves are required together — half a cursor pages
        from an instant with no tie-break, which is what the id exists to prevent.
        Offset paging over a table that grows at the head repeats and skips rows,
        which on a ledger reads as the log lying.

        ⚠️ The ledger is PRUNED (see `retentionDays` in the response). Asking for a
        window older than that returns nothing because the rows are gone, not
        because nothing happened.
      parameters:
        - { name: sessionId, in: query, schema: { type: string } }
        - { name: endpoint, in: query, schema: { type: string }, description: One of the values in the response's `endpoints`. }
        - { name: model, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [ok, error] } }
        - { name: since, in: query, schema: { type: string, format: date-time }, description: RFC3339 lower bound, inclusive. }
        - { name: beforeAt, in: query, schema: { type: string }, description: 'Cursor half; requires beforeId.' }
        - { name: beforeId, in: query, schema: { type: string }, description: 'Cursor half; requires beforeAt.' }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs: { type: array, items: { $ref: '#/components/schemas/AICallLog' } }
                  endpoints: { type: array, items: { type: string }, description: This build's endpoint vocabulary, for the filter. }
                  retentionDays: { type: integer }
                  nextBeforeAt: { type: string, description: Absent when this page did not fill. }
                  nextBeforeId: { type: string }
        '400': { description: Bad status, since, or half a cursor. }
        '403': { $ref: '#/components/responses/Error' }
  /api/v2/admin/usage:
    get:
      operationId: adminUsage
      tags: [admin]
      security: [{ adminToken: [] }]
      summary: Model spend history — per day and per model, kept forever.
      description: >-
        Read from `ai_usage_daily`, a rollup folded server-side (one
        `INSERT … SELECT … GROUP BY` per closed day on SQL, one aggregation
        pipeline ending in `$merge` on Mongo). The raw ledger is pruned at ninety
        days; this is not, so the figures are history rather than a window.

        ⚠️ TODAY IS NOT INCLUDED, in either fold. A day is folded only once it can
        no longer receive rows, so the rollup stops at `throughDay`.
        `GET /api/admin/stats` is the one place that adds today's live ledger rows
        on top — one implementation of "history plus today" rather than two that
        can drift.

        Two folds in one response because neither is actionable alone: the day
        series says spend doubled, the model fold says which model did it.

        ⚠️ Same permission as the ledger (ailogs.read), deliberately NOT
        overview.read — a cost breakdown is not "is it up", and overview.read is
        the permission everybody gets.

        Rows in `byModel` carry no `day`; they are a fold ACROSS days.
      parameters:
        - { name: from, in: query, schema: { type: string }, description: 'YYYY-MM-DD, UTC, inclusive.' }
        - { name: to, in: query, schema: { type: string }, description: 'YYYY-MM-DD, UTC, inclusive.' }
        - { name: limit, in: query, schema: { type: integer, default: 90, maximum: 800 }, description: Days in the series. }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals: { $ref: '#/components/schemas/AIUsageTotals' }
                  days: { type: array, items: { $ref: '#/components/schemas/AIUsageDay' } }
                  byModel: { type: array, items: { $ref: '#/components/schemas/AIUsageDay' } }
                  throughDay: { type: string, description: Newest day the rollup can cover. }
        '400': { description: from or to is not YYYY-MM-DD. }
        '403': { $ref: '#/components/responses/Error' }
  /api/v2/admin/users:
    get:
      operationId: adminUsers
      tags: [admin]
      security: [{ adminToken: [] }]
      summary: Who exists, who owns this deployment, who is in what, and who is using how much.
      description: >-
        ⚠️ Counts, never content. users.read promises "不含任何人写下的内容" — how
        many calls somebody made is an operational fact about the deployment; what
        they said is not, and lives behind db.user_content.

        `fastWindowHours` / `slowWindowDays` are the server's own window lengths,
        returned so a screen labels the counters with the numbers the server
        actually resets on rather than a copy that drifts.
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        email: { type: string }
                        name: { type: string }
                        owner: { type: boolean }
                        roles: { type: array, items: { type: string } }
                        createdAt: { type: string, format: date-time }
                        usage: { $ref: '#/components/schemas/SessionUsage' }
                  ownerCount: { type: integer }
                  limit: { type: integer }
                  fastWindowHours: { type: integer }
                  slowWindowDays: { type: integer }
        '403': { $ref: '#/components/responses/Error' }
  /api/v2/admin/users/{id}:
    delete: { operationId: adminDeleteUser, tags: [admin], security: [{ adminToken: [] }], parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/Error' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: dc_sid }
    authCookie: { type: apiKey, in: cookie, name: dc_auth }
    sessionToken: { type: apiKey, in: header, name: X-Session-Token, description: "Same signed value as the dc_sid cookie, for clients without a cookie jar. Get it from POST /api/session/init {tokenInBody:true}." }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: "Same JWT as the dc_auth cookie. Get it from login/register {tokenInBody:true}; logout revokes all issued JWTs (token-version bump)." }
    importToken: { type: apiKey, in: header, name: X-Import-Token, description: "Session-scoped token for the browser extension's direct push." }
    adminToken: { type: apiKey, in: header, name: X-Admin-Token, description: "Admin endpoints." }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: "Stable code the frontend translates (e.g. no_session, bad_request, rate_limited, token_expired)." }
        message: { type: string, description: "Localized fallback message." }
...
    AICallLog:
      type: object
      description: One model call. Metadata only — see the ailogs endpoint for why there is no body here.
      properties:
        id: { type: string }
        sessionId: { type: string }
        endpoint: { type: string }
        model: { type: string }
        promptTokens: { type: integer }
        compTokens: { type: integer }
        durationMs: { type: integer }
        status: { type: string, enum: [ok, error] }
        error: { type: string, description: Truncated to 300 characters. }
        requestId: { type: string }
        createdAt: { type: string, format: date-time }
    AIUsageDay:
      type: object
      description: >-
        One fold of the spend rollup. In the day series `model` and `endpoint`
        are empty (summed across them); in the model fold `day` is empty
        (summed across days).
      properties:
        day: { type: string, description: 'YYYY-MM-DD, UTC.' }
        model: { type: string }
        endpoint: { type: string }
        calls: { type: integer }
        errors: { type: integer }
        promptTokens: { type: integer }
        compTokens: { type: integer }
    AIUsageTotals:
      type: object
      properties:
        calls: { type: integer }
        errors: { type: integer }
        promptTokens: { type: integer }
        compTokens: { type: integer }
        firstDay:
          type: string
          description: >-
            The oldest day the rollup actually holds. ⚠️ Render it. The rollup
            can only be built from ledger rows that still exist, so on a
            deployment upgrading into this feature the total begins where the
            ledger began — not at the beginning of time. A figure shown without
            this date is the same lie the rollup exists to fix, one level up.
    SessionUsage:
      type: object
      description: >-
        One account's AI spend at three scales, stored on the session row rather
        than in a table — nine columns and zero new rows, against a (users ×
        days) table that grows forever to answer a question almost always asked
        about NOW.

        ⚠️ The windows TUMBLE, they do not slide: when a window expires the next
        call RESETS its counter rather than dropping the oldest events. So just
        after a rollover a busy account reads small. A true sliding window needs
        per-event rows, which is the table this avoids. It is the right
        approximation for what these are for — a quota that resets is what a
        quota IS.

        ⚠️ An expired window is NOT zeroed in storage; nothing runs on expiry.
        Callers get it already zeroed here, but anything reading the raw row must
        check the start stamp itself.

        ⚠️ This is a HOOK, not billing. Per-model attribution, an immutable
        record per period, and a link to an ACCOUNT rather than a session are all
        deliberately absent — see docs/ROADMAP.md.
      properties:
        fastCalls: { type: integer }
        fastTokens: { type: integer, description: Prompt and completion combined. }
        fastWindowStart: { type: string, format: date-time }
        slowCalls: { type: integer }
        slowTokens: { type: integer }
        slowWindowStart: { type: string, format: date-time }
        totalCalls: { type: integer, description: Lifetime; never reset. }
        totalPromptTokens: { type: integer }
        totalCompTokens: { type: integer }
...
    AdminStats:
      type: object
      properties:
        users: { type: integer }
        sessions: { type: integer }
        aiCalls: { type: integer }
        tokenUsed: { type: integer }
        feedbackUseful: { type: integer }
        feedbackTotal: { type: integer }
...